refactor: Relocate all information_schema tests into a single test file.
